third_party/blink/web_tests/external/wpt/speculation-rules/speculation-tags/deduped-and-sorted-tags.https.html

<!DOCTYPE html>
<title>Sec-Speculation-Tags request headers</title>
<meta name="variant" content="?type=prefetch">
<meta name="variant" content="?type=prerender">
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="/common/utils.js"></script>
<script src="/html/browsers/browsing-the-web/remote-context-helper/resources/remote-context-helper.js"></script>
<script src="/speculation-rules/resources/utils.js"></script>
<script src="resources/speculation-tags-utils.js"></script>
<script>
"use strict";

setup(() => assertSpeculationRulesIsSupported());

function constructSpeculationRule(preloadingType, url, rulesetTag, ruleTags) {
  // Construct rules.
  const rules = [];
  for (const tag of ruleTags) {
    rules.push({tag, source: "list", urls: [url]});
  }

  // Construct the ruleset.
  return JSON.stringify({
      tag: rulesetTag,
      [preloadingType]: rules
  });
}

// Test the case where multiple tags (including duplicate tags, invalid tags,
// and no tags) specified on both the ruleset level and the rule level. These
// should be deduped and sorted before sent as the header.
promise_test(async t => {
    const params = new URLSearchParams(window.location.search);
    const preloadingType = params.get('type');

    // Inputs and expectations:
    // "def" is duplicate. "\x19" is invalid. "undefined" indicates no tag.
    const rulesetTag = "def";
    const ruleTags = ["def", "jkl", "def", "null", "\x19", "abc", undefined, "ghi"];
    // "def" should be deduped.
    // "\x19" and "undefined" should be ignored (should not send null tag).
    const expectedTags = '"abc", "def", "ghi", "jkl", "null"';

    const rcHelper = new RemoteContextHelper();
    const referrerRC = await rcHelper.addWindow(undefined, { features: 'noopener' });

    const extraConfig = {};
    const preloadedRC = await referrerRC.helper.createContext({
        executorCreator(url) {
          const speculationRule = constructSpeculationRule(
              preloadingType, url, rulesetTag, ruleTags);
          return referrerRC.executeScript((speculationRule) => {
              const script = document.createElement("script");
              script.type = "speculationrules";
              script.textContent = speculationRule;
              document.head.append(script);
          }, [speculationRule]);
        }, extraConfig
    });

    // Navigate to the preloaded page.
    await referrerRC.navigateTo(preloadedRC.url);

    const headers = await preloadedRC.getRequestHeaders();
    // Make sure the page is preloaded.
    assert_equals(
      headers.get("sec-purpose"),
      preloadingType === "prefetch" ? "prefetch" : "prefetch;prerender");
    assert_equals(headers.get("sec-speculation-tags"), expectedTags);
}, "Sec-Speculation-Tags: deduped and sorted tags");

// Test the case where multiple tags (including duplicate tags, invalid tags,
// and no tags) specified only on the rule level. These should be deduped and
// sorted before sent as the header. Unlike the previous test, no tag on the
// rule should result in the null tag.
promise_test(async t => {
    const params = new URLSearchParams(window.location.search);
    const preloadingType = params.get('type');

    // Inputs and expectations:
    // No ruleset level tag.
    const rulesetTag = undefined;
    // "def" is duplicate. "\x19" is invalid. "undefined" indicates no tag.
    const ruleTags = ["def", "jkl", "def", "null", "\x19", "abc", undefined, "ghi"];
    // 5 rules are (no ruleset level tag) + (rule level tag) =
    //     "abc" only, "def" only, "ghi" only, "jkl" only, "null" only.
    // 1 rule is (no ruleset level tag) + (no rule level tag) = null only.
    // 2 rule are duplicate or invalid.
    //
    // All 6 effective rules, including the null rule, get combined to match
    // expectedTags.
    const expectedTags = 'null, "abc", "def", "ghi", "jkl", "null"';

    const rcHelper = new RemoteContextHelper();
    const referrerRC = await rcHelper.addWindow(undefined, { features: 'noopener' });

    const extraConfig = {};
    const preloadedRC = await referrerRC.helper.createContext({
        executorCreator(url) {
          const speculationRule = constructSpeculationRule(
              preloadingType, url, rulesetTag, ruleTags);
          return referrerRC.executeScript((speculationRule) => {
              const script = document.createElement("script");
              script.type = "speculationrules";
              script.textContent = speculationRule;
              document.head.append(script);
          }, [speculationRule]);
        }, extraConfig
    });

    // Navigate to the preloaded page.
    await referrerRC.navigateTo(preloadedRC.url);

    const headers = await preloadedRC.getRequestHeaders();
    // Make sure the page is preloaded.
    assert_equals(
      headers.get("sec-purpose"),
      preloadingType === "prefetch" ? "prefetch" : "prefetch;prerender");
    assert_equals(headers.get("sec-speculation-tags"), expectedTags);
}, "Sec-Speculation-Tags: deduped and sorted tags within rules");

</script>
